+++ objects/common.h
@@ -13,7 +13,7 @@ class CommonObject {
 
     double getDistance(sf::Vector2i a, sf::Vector2i b);
     double getAngle(sf::Vector2i a, sf::Vector2i b);
-    std::string padIntToString(int s, unsigned int sz, char p);
+    std::string padIntToString(long s, unsigned int sz, char p);
 
   public:
     virtual void update(const sf::Time &t) = 0;
