+++ states/tools/ultrasoundtool.cpp
@@ -11,8 +11,8 @@ void UltrasoundToolState::initRipple() {
 
 void UltrasoundToolState::updateRipples(const sf::Time &time) {
   // update ongoing ripples, check for expired
-  std::vector<unsigned int> expired;
-  for (unsigned int i = 0; i < ripples.size(); i++) {
+  std::vector<unsigned long> expired;
+  for (unsigned long i = 0; i < ripples.size(); i++) {
     if (ripples[i]->hasExpired()) {
       expired.push_back(i);
     } else {
@@ -20,8 +20,8 @@ void UltrasoundToolState::updateRipples(const sf::Time &time) {
     }
   }
   // removed expired ripples, if any
-  for (unsigned int i = expired.size(); i > 0; i--) {
-    unsigned int index = expired[i-1];
+  for (unsigned long i = expired.size(); i > 0; i--) {
+    unsigned long index = expired[i-1];
     delete ripples[index];
     ripples.erase(ripples.begin() + index);
   }
@@ -50,7 +50,7 @@ void UltrasoundToolState::update(const sf::Time &time, bool active) {
 
 void UltrasoundToolState::draw(sf::RenderWindow *window, bool active) {
   // draw ongoing ripples
-  for (unsigned int i = 0; i < ripples.size(); i++) {
+  for (unsigned long i = 0; i < ripples.size(); i++) {
     ripples[i]->draw(window);
   }
 }
@@ -61,7 +61,7 @@ void UltrasoundToolState::deselect() {
 
 UltrasoundToolState::~UltrasoundToolState() {
   deselect();
-  for (unsigned int i = 0; i < ripples.size(); i++) {
+  for (unsigned long i = 0; i < ripples.size(); i++) {
     delete ripples[i];
   }
 }
